refactor: moved interface to declatation file
