+++ src/vite-env.d.ts
@@ -1,6 +1,7 @@
 /// <reference types="vite/client" />
 
 import 'vue-router';
+import 'jwt-decode';
 
 /*
  * Declare module '*.vue' {
@@ -23,5 +24,19 @@ declare module '*.vue' {
 declare module 'vue-router' {
 	interface RouteMeta {
 		requiresAuthRole?: Express.User['role'][]
+	}
+}
+
+declare module 'jwt-decode' {
+	interface JwtPayload {
+		iss?: string;
+		sub?: string;
+		aud?: string[] | string;
+		exp?: number;
+		nbf?: number;
+		iat?: number;
+		jti?: string;
+		role?: 'User' | 'Admin';
+		userId?: string;
 	}
 }
